Write book's actions and reducer
