+++ src/pages/Categories.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Categories() {
+  return (
+    <div>
+      <button type="button">Check status</button>
+    </div>
+  );
+}
+
+export default Categories;
