--- src/redux/categories/Categories.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Categories() {
-  return (
-    <div>
-      <button type="button">Check status</button>
-    </div>
-  );
-}
-
-export default Categories;
